+++ docs/CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 =========
-- Enh: Add Create Button
+- Enh #423: Add Create Button
 
 1.5.3 (September 14, 2023)
 --------------------------