+++ docs/CHANGELOG.md
@@ -1,10 +1,10 @@
 Changelog
 =========
-- Enh #423: Add Create Button
 
 1.5.4 - (Unreleased)
 -----------------------
 - Enh #424: Tests for `next` version
+- Enh #423: Add Create Button
 
 1.5.3 (September 14, 2023)
 --------------------------