+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 - Enh #424: Tests for `next` version
 - Enh #423: Add Create Button
+- Fix #419: Don't use loader for wall entry menu on open modal window 
 
 1.5.3 (September 14, 2023)
 --------------------------