+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 - Enh #424: Tests for `next` version
 - Enh #423: Add Create Button
+- Fix #419: Don't use loader for wall entry menu on open modal window 
 - Enh #379: Allow managing of global event types from container side
 - Enh #382: Allow inviting by users who can manage a calendar entry